django_crud/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
distlib==0.3.9
Django==5.1.3
django-environ==0.11.2
django-pagination==1.0.7
django-unfold==0.42.0
Faker==33.1.0
filelock==3.16.1
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
virtualenv==20.28.0
